IRanges 1.7.24
C-level constructors alloc_XRawList(), alloc_XIntegerList(),
alloc_XDoubleList() and new_XRawList_from_CharAEAE() are now
able to produce an XVectorList object of "arbitrary" size i.e.
an object with <= 2^31-1 elements of length <= 2^31-1 each
(there used to be the additional restriction that the sum of the
lengths of the elements must be <= 2^31-1).
This improvement directly benefits Biostrings functions
read.DNAStringSet() (and family), extractTranscripts() and translate().


git-svn-id: file:///home/git/hedgehog.fhcrc.org/bioconductor/trunk/madman/Rpacks/IRanges@49003 bc3139a8-67e5-0310-9ffc-ced21a209358