IPos objects now exist in 2 flavors: UnstitchedIPos and StitchedIPos
IPos is now a virtual class with 2 concrete subclasses: UnstitchedIPos
and StitchedIPos. In an UnstitchedIPos instance the positions are stored
as an integer vector. In a StitchedIPos instance, like with old IPos
instances, the positions are stored as an IRanges object where each range
represents a run of consecutive positions. See ?IPos for more information.

Old serialized IPos instances will need to be converted to StitchedIPos
instances with updateObject() (not ready yet).