RangedData objects are now defunct
RangedData objects are defunct in BioC 3.11. They were deprecated in BioC
3.9 and, before that, their used has been discouraged in favor of GRanges
or GRangesList objects since BioC 2.12, that is, since 2014.
Maintainers of packages that still use RangedData objects need to migrate
their code to use GRanges or GRangesList objects as soon as possible (the
deprecation message in BioC 3.9 and BioC 3.10 asked them to do so for about
a year already).