coverage() now supports 'method="naive"'
This is in addition to the already supported methods "sort" and "hash".
This new method is a slower version of the "hash" method that has the
advantage of avoiding floating point artefacts in the no-coverage regions
of the numeric-Rle object returned by coverage() when the weights are
supplied as a numeric vector of type 'double'.